+++ RELEASE_NOTES.md
@@ -1,7 +1,7 @@
 # Release Notes
 
 
-## 1.4
+## 0.4
 
 - This version localizes all texts.